requirements.txt
streamlit
langchain
langchain_core
langchain_community
langchain_openai
langgraph
requests
beautifulsoup4
python-dotenv
tavily-python